requirements.txt
torch~=2.3.1
torchaudio~=2.3.1
tqdm~=4.66.4
scipy~=1.14.0
pystoi~=0.4.1
pesq~=0.0.4
numpy~=2.0.1
python-dotenv~=1.0.1
metrics~=0.3.3
tensorflow~=2.16.2
soundfile~=0.12.1
numba~=0.60.0